Moved SyncDateWrapper to org.hisp.dhis.client.sdk.ui.bindings.commons.SyncDateWrapper
