+++ ui-bindings/src/main/java/org/hisp/dhis/client/sdk/ui/bindings/commons/DefaultAppModuleImpl.java
@@ -7,7 +7,6 @@ import org.hisp.dhis.client.sdk.android.api.D2;
 import org.hisp.dhis.client.sdk.android.api.utils.LoggerImpl;
 import org.hisp.dhis.client.sdk.ui.AppPreferences;
 import org.hisp.dhis.client.sdk.ui.AppPreferencesImpl;
-import org.hisp.dhis.client.sdk.ui.SyncDateWrapper;
 import org.hisp.dhis.client.sdk.utils.Logger;
 
 import static org.hisp.dhis.client.sdk.utils.Preconditions.isNull;
