+++ ui-bindings/src/main/java/org/hisp/dhis/client/sdk/ui/bindings/commons/UserComponent.java
@@ -4,7 +4,6 @@ import android.content.Context;
 
 import org.hisp.dhis.client.sdk.android.user.CurrentUserInteractor;
 import org.hisp.dhis.client.sdk.ui.AppPreferences;
-import org.hisp.dhis.client.sdk.ui.SyncDateWrapper;
 import org.hisp.dhis.client.sdk.ui.bindings.presenters.HomePresenter;
 import org.hisp.dhis.client.sdk.ui.bindings.presenters.LauncherPresenter;
 import org.hisp.dhis.client.sdk.ui.bindings.presenters.LoginPresenter;
