+++ ui-bindings/src/main/java/org/hisp/dhis/client/sdk/ui/bindings/presenters/ProfilePresenterImpl.java
@@ -2,7 +2,7 @@ package org.hisp.dhis.client.sdk.ui.bindings.presenters;
 
 import org.hisp.dhis.client.sdk.android.user.CurrentUserInteractor;
 import org.hisp.dhis.client.sdk.models.user.UserAccount;
-import org.hisp.dhis.client.sdk.ui.SyncDateWrapper;
+import org.hisp.dhis.client.sdk.ui.bindings.commons.SyncDateWrapper;
 import org.hisp.dhis.client.sdk.ui.bindings.commons.DefaultAppAccountManager;
 import org.hisp.dhis.client.sdk.ui.bindings.commons.DefaultNotificationHandler;
 import org.hisp.dhis.client.sdk.ui.bindings.commons.RxOnValueChangedListener;
